Factor out BuildStatsData.py for reuse (#7376)
Will reuse this in gather_build_stats.py

Other small refactoring and cleanup.
